crypto_agent_v1/requirements.txt
langgraph
langchain-ollama
ccxt
pandas
pandas-ta
pyyaml
requests
